fix(dyte-leave-meeting): use consistent primary button in leave dialog
For users who sometimes (but don't always) have permission to end a
meeting for everyone, the big red button was doing different things,
which could potentially lead someone unintentionally ending the meeting
for everyone when they were only trying to leave the meeting. By using
consistent "primary" styling for the "Leave" button, we can prevent
this, and de-emphasize the most destructive action.